requirements.txt
Telethon==1.24.0
telethon-tgcrypto
loguru==0.6.0
python-dotenv==0.19.2
pytz==2021.3
psutil==5.9.0
heroku3==5.1.4
GitPython==3.1.27
gpytranslate==1.4.0
gTTS==2.2.3
aiofiles==0.8.0
aiocsv==1.2.1
aiohttp[speedups]==3.8.1
beautifulsoup4==4.10.0
Markdown==3.3.6
emoji==1.6.3
